feat: move CurrentReleaseVersion const to Chainflip
