+++ state-chain/pallets/cf-environment/src/mock.rs
@@ -133,14 +133,6 @@ impl VaultKeyWitnessedHandler<Bitcoin> for MockBitcoinVaultKeyWitnessedHandler {
 	}
 }
 
-parameter_types! {
-	pub CurrentReleaseVersion: SemVer = SemVer {
-		major: env!("CARGO_PKG_VERSION_MAJOR").parse::<u8>().unwrap(),
-		minor: env!("CARGO_PKG_VERSION_MINOR").parse::<u8>().unwrap(),
-		patch: env!("CARGO_PKG_VERSION_PATCH").parse::<u8>().unwrap(),
-	};
-}
-
 pub struct MockBitcoinFeeInfo;
 impl GetBitcoinFeeInfo for MockBitcoinFeeInfo {
 	fn bitcoin_fee_info() -> BitcoinFeeInfo {
@@ -161,7 +153,6 @@ impl pallet_cf_environment::Config for Test {
 	type BitcoinVaultKeyWitnessedHandler = MockBitcoinVaultKeyWitnessedHandler;
 	type BitcoinFeeInfo = MockBitcoinFeeInfo;
 	type RuntimeSafeMode = MockRuntimeSafeMode;
-	type CurrentReleaseVersion = CurrentReleaseVersion;
 	type WeightInfo = ();
 }
